requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
telebot==0.0.5